List containers API call to list containers on current node
